Deprecated EnsureWeb and EnsureSite methods
